+layout.load で context を先読み
何度も context を初期化してしまうのを回避
ただ、タイマーを映さないページでも初期化が走ってしまう
PiP をつねに維持するためと考えれば、別に悪くないか？
